% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencyTimings-package.R
\docType{package}
\name{dependencyTimings-package}
\alias{dependencyTimings}
\alias{dependencyTimings-package}
\title{dependencyTimings: Time combinations of loading packages}
\description{
To determine which package dependencies might be contributing to long loading
    times, profiles loading times of package dependencies in independent R sessions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rmflight/dependencyTimings}
  \item Report bugs at \url{https://github.com/rmflight/dependencyTimings/issues}
}

}
\author{
\strong{Maintainer}: Robert M Flight \email{rflight79@gmail.com}

}
\keyword{internal}
