% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{get_imports}
\alias{get_imports}
\title{get imported packages}
\usage{
get_imports(package, suggests = FALSE)
}
\arguments{
\item{package}{the name of the package for the imports}

\item{suggests}{should the packages in suggests be returned (default = FALSE)}
}
\value{
character
}
\description{
given a package name, extracts everything from the \emph{Imports} and \emph{Depends}
fields.
}
