% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{get_package_imports}
\alias{get_package_imports}
\title{get package imports}
\usage{
get_package_imports(package)
}
\arguments{
\item{package}{the name of the package for the imports}
}
\value{
character
}
\description{
given a package name, extracts everything from the imports field.
}
