% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencyTimings-package.R
\docType{package}
\name{importedPackageTimings-package}
\alias{importedPackageTimings}
\alias{importedPackageTimings-package}
\title{importedPackageTimings: Benchmark Load Timings of Imported Packages}
\description{
To determine which package dependencies (imports) might be contributing to long loading
    times, profiles loading times of package imports in independent R sessions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rmflight/dependencyTimings}
  \item Report bugs at \url{https://github.com/rmflight/dependencyTimings/issues}
}

}
\author{
\strong{Maintainer}: Robert M Flight \email{rflight79@gmail.com}

}
\keyword{internal}
