% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcms_time.R
\docType{data}
\name{xcms_time}
\alias{xcms_time}
\title{Timings for xcms dependencies}
\format{A data frame with 26 rows and 7 variables:
\describe{
\item{package}{which package it is}
\item{med}{median timings in microseconds}
\item{min}{minimum load time in microseconds}
\item{max}{maximum load time}
\item{type}{load time for the dependency (pkg), or for the package after loading the dependency (after)}
\item{which}{the package or a dependency}
\item{timings}{the raw times for each replicate}
}}
\source{
Rober M Flight on personal machine
}
\usage{
xcms_time
}
\description{
A dataset containing the \code{library(pkg)} timings for the dependencies of
the \code{xcms} package.
}
\keyword{datasets}
